committed

+++ src/main/java/com/sunglowsys/repository/RatePlanRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RatePlanRepository extends JpaRepository<RatePlan, Integer> {
+public interface RatePlanRepository extends JpaRepository<RatePlan, Long> {
 
 }
